Replace spaces with dashes in a folder name
In the GitHub iOS app, if set to open external links in some browsers
including Brave and Firefox (but not Safari or DuckDuckGo), each `%20`
(representing a space) in a link's URL would be replaced with `%2520`,
breaking the link. This commit replaces spaces to avoid this encoding
bug that some mobile browsers appear to have.